requirements.txt
celery==5.0.2
Django==3.1.3
django-cors-headers==3.6.0
djangorestframework==3.12.2
redis==3.5.3
python-magic==0.4.13
